requirements.txt
colorama==0.4.3
kafka-python==2.0.1
